Refactoring : infer
